Split main thread webview views to own file
